Move getVarFieldContent into the marc namespace
